Fix attribute adding to not duplicate attributes
Previously, the plugin would add a new attribute to the relevant
markdown-it token for each that it encountered. However, if the
attribute already existed, this duplicated the attribute. Because of the
way that markdown-it reads attributes, the duplicate is essentially
ignored. By using `Token.attrSet` instead of `Token.attrPush`,
markdown-it will automatically de-duplicate attributes for us.

See https://markdown-it.github.io/markdown-it/#Token.attrSet.